% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_frag_info.R
\name{add_frag_info}
\alias{add_frag_info}
\title{decode fragment identifiers for spike-in standards}
\usage{
add_frag_info(x, frag_grp = "frag_grp")
}
\arguments{
\item{x}{data.frame with a column of frag_grp information (see above)}

\item{frag_grp}{the column name for the frag_grp information (\code{frag_grp})}
}
\value{
\preformatted{     the data.frame x, augmented with fraglen, CpG, and GC cols
}
}
\description{
given a vector of fragment identifiers like \verb{160_2_35}, encoded as
lengthInBp_numberOfCpGs_GCpercent, pull them apart into a data.frame,
add those columns to the source data.frame (whence \code{frag_grp} or similar),
and return the updated data.frame.
}
\examples{
data(spike_read_counts) 
add_frag_info(spike_read_counts)

}
