% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedpe_covg.R
\name{bedpe_covg}
\alias{bedpe_covg}
\title{get coverage from a BEDPE file (possibly removing duplicated bits)}
\usage{
bedpe_covg(x, param = NULL, stranded = FALSE, undup = TRUE)
}
\arguments{
\item{x}{a Tabixed BEDPE file, or a TabixFile of one}

\item{param}{a GRanges of regions to read, or NULL (NULL)}

\item{stranded}{treat pairs as stranded? (FALSE)}

\item{undup}{remove overlapping bits? (TRUE)}
}
\value{
\preformatted{     an RleList of coverage
}
}
\description{
get coverage from a BEDPE file (possibly removing duplicated bits)
}
