% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pmol.R
\name{bedtools_wao_imitation}
\alias{bedtools_wao_imitation}
\title{union-intersect-overlap for things with values}
\usage{
bedtools_wao_imitation(A, B)
}
\arguments{
\item{A}{the ranges spanned by fragments (see Details)}

\item{B}{the bins (typically 300bp wide) to tally in (see Details)}
}
\value{
\preformatted{ a GRanges with disjoint ranges and counts of A in each B 
}
}
\description{
like it says on the tin: pretend to do \verb{bedtools -wao -a A -b B}
this should be preceded by B <- bam_to_bins(bam) to work as expected.
}
