% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPairedGRtoGR.R
\name{convertPairedGRtoGR}
\alias{convertPairedGRtoGR}
\title{Convert Pairs to GRanges}
\usage{
convertPairedGRtoGR(pairs)
}
\arguments{
\item{pairs}{the Pairs object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       a GRanges
}\if{html}{\out{</div>}}
}
\description{
Convert Pairs to GRanges
}
