% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dedup}
\alias{dedup}
\title{spike-in counts for two samples, as a wide data.frame}
\format{
A data.frame object with
\describe{
\item{frag_grp}{the encoded spike contig name: basepairs_CpGs_GCpercent}
\item{read_count_6547}{read coverage for this spike in sample 6547}
\item{read_count_6548}{read coverage for this spike in sample 6548}
}
}
\usage{
dedup
}
\description{
A data.frame with spike-in results from control samples in the manuscript.
This maps 1:1 onto \code{spike_read_counts} using reshape2::melt.
}
\keyword{datasets}
