% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spike_contigs.R
\name{find_spike_contigs}
\alias{find_spike_contigs}
\title{find spike-in seqlevels in an object \code{x}, where !is.null(seqinfo(x))}
\usage{
find_spike_contigs(x, spike = NULL)
}
\arguments{
\item{x}{something with seqlevels}

\item{spike}{a DataFrame with spike-in information, if not using defaults}
}
\value{
\preformatted{   indices of which contigs in seqlevels(x) are spike-in contigs
}
}
\description{
Find the spike-like contigs in a BAM with both natural and spiked contigs.
This started out as glue in some other functions and got refactored out.
}
\details{
The indices have an attribute "mappings", which is a character vector
such that attr(find_spike_contigs(x), "mappings")\link{original} == standardized
where original is the name in the CRAM/BAM/whatever, and standardized is
the rowname in \code{spike} that corresponds to that original contig name.
}
\seealso{
get_base_name

rename_spike_seqlevels
}
