% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_coverage.R
\name{format_coverage}
\alias{format_coverage}
\title{take an Rle\link{List} and either leave it alone (ret="rle") or transform it}
\usage{
format_coverage(x, contigs, param = NULL, ret = c("rle", "df"))
}
\arguments{
\item{x}{the BAM or CRAM (or BEDPE)}

\item{contigs}{the contigs to format}

\item{param}{a ScanBamParam object (ignored if BEDPE)}

\item{ret}{RleList ("rle") or data.frame ("df")}
}
\value{
\preformatted{     depends on the value of `ret`
}
}
\description{
Since a lot of spiky functions have to summarize coverage, this shims that.
}
