% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genbank_mito}
\alias{genbank_mito}
\title{various mitochondrial genomes sometimes used as endogenous spike-ins}
\format{
A DataFrame object with
\describe{
\item{species}{the species whence the record came, as a character string}
\item{genome}{the genome assembly whence the mtDNA, as a character string}
\item{accession}{the genbank accession, as a character string}
\item{sequence}{genome sequence, as a DNAStringSet}
}
}
\usage{
data(genbank_mito)
}
\description{
A DataFrame with species, genome, accession, and sequence for GenBank
mitochondrial genome depositions. No concentration provided; add if needed.
}
\keyword{datasets}
