% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spike_fasta.R
\name{generate_spike_fasta}
\alias{generate_spike_fasta}
\title{for CRAM files, a FASTA reference is required to decode; this builds that}
\usage{
generate_spike_fasta(x, fasta = "spike_contigs.fa", spike = NULL)
}
\arguments{
\item{x}{a CRAM file, hopefully with an index}

\item{fasta}{the filename for the resulting FASTA ("spikes.fa")}

\item{spike}{a DataFrame where spike$sequence is a DNAStringSet (spike)}
}
\value{
\preformatted{     a DNAStringSet with renamed contigs, as exported to `fasta` 
}
}
\description{
If the contigs in a CRAM have even slightly different names from those in
the reference, decoding will fail.  In some cases there are multiple names
for a given contig (which raises the question of whether to condense them),
and thus the same reference sequence decodes multiple contig names.
}
\details{
This function simply generates an appropriate spike reference for a CRAM,
using the CRAM headers to figure out which references are used for which.
At the moment, CRAM support in Rsamtools only exists in the GitHub branch:

BiocManager::install("Bioconductor/Rsamtools@cram")

Using other versions of Rsamtools will yield an error.
}
