% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genomic_res}
\alias{genomic_res}
\title{A Granges object with genomic coverage from chr21q22,
binned every 300bp for the genomic contigs then averaged across the bin. (In other
words, the default output of scan_genomic_contigs or scan_genomic_bedpe,
restricted to a small enough set of genomic regions to be practical for examples.)
This represents what most users will want to generate from their own genomic BAMs or BEDPEs, and
is used repeatedly in downstream examples throughout the package.}
\format{
A GRanges of coverage results with one metadata column, \code{coverage}
}
\source{
Generated using scan_genomic_bedpe or scan_genomic_contigs on an example bedpe or bam
containing chr21q22 contigs.
}
\usage{
data(genomic_res)
}
\description{
A Granges object with genomic coverage from chr21q22,
binned every 300bp for the genomic contigs then averaged across the bin. (In other
words, the default output of scan_genomic_contigs or scan_genomic_bedpe,
restricted to a small enough set of genomic regions to be practical for examples.)
This represents what most users will want to generate from their own genomic BAMs or BEDPEs, and
is used repeatedly in downstream examples throughout the package.
}
\keyword{datasets}
