% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spike_depth.R
\name{get_spike_depth}
\alias{get_spike_depth}
\title{get the (max, median, or mean) coverage for spike-in contigs from a BAM/CRAM}
\usage{
get_spike_depth(covg, spike_gr, how = c("max", "mean"), spike = NULL)
}
\arguments{
\item{covg}{the coverage RleList}

\item{spike_gr}{the spike-in GRanges}

\item{how}{how to summarize the per-spike coverage (max)}

\item{spike}{information about the spikes (default: load \code{spike})}
}
\value{
\preformatted{     a GRanges with summarized coverage and features for each
}
}
\description{
get the (max, median, or mean) coverage for spike-in contigs from a BAM/CRAM
}
\examples{
sb <- system.file("extdata", "example.spike.bam", package="spiky", 
                  mustWork=TRUE)
si <- seqinfo_from_header(sb) 
genome(si) <- "spike"
mgr <- get_merged_gr(si)

fl <- scanBamFlag(isDuplicate=FALSE, isPaired=TRUE, isProperPair=TRUE)
bp <- ScanBamParam(flag=fl)
bamMapqFilter(bp) <- 20

covg <- get_spiked_coverage(sb, bp=bp, gr=mgr)
get_spike_depth(covg, spike_gr=mgr)

}
