% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spiked_coverage.R
\name{get_spiked_coverage}
\alias{get_spiked_coverage}
\title{tabulate coverage across assembly and spike contig subset in natural order}
\usage{
get_spiked_coverage(bf, bp, gr)
}
\arguments{
\item{bf}{the BamFile object}

\item{bp}{the ScanBamParam object}

\item{gr}{the GRanges with sorted seqlevels}
}
\value{
a list of Rles
}
\description{
Refactored from scan_spiked_bam, this is a very simple wrapper
}
\seealso{
scan_spiked_bam

coverage
}
