% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarySize.R
\name{librarySize}
\alias{librarySize}
\title{adapted from code in Caleb Lareau's mtscATAC paper, which adapts Picard's}
\usage{
librarySize(nT, nU)
}
\arguments{
\item{nT}{total number of genomic fragments}

\item{nU}{unique genomic fragments}
}
\value{
\preformatted{   estimated number of fragments in the source library
}
}
\description{
adapted from code in Caleb Lareau's mtscATAC paper, which adapts Picard's
}
