% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_glm_pmol.R
\name{model_glm_pmol}
\alias{model_glm_pmol}
\title{Attempts to build a generalized linear model from spike-in standards,
to correct bias and batch effects in a cell-free MeDIP experiment.}
\usage{
model_glm_pmol(x, concentration = NULL, ...)
}
\arguments{
\item{x}{the data, with (at minimum) frag_grp, id, read_count}
}
\value{
\preformatted{   the model fit for the data
}
}
\description{
formerly '2020_model_glm_fmol'.  Note that everything in X can be had from
a BAM/CRAM with spike contigs named as frag_grp (len_CpGs_GC) in the index
}
\details{
\preformatted{  The format of 'data' is important, yet poorly documented
           apparently column 2 is fragment length as a character 
           somewhere we find the number of CpGs as an integer
}
}
