% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spike_UMI.R
\name{parse_spike_UMI}
\alias{parse_spike_UMI}
\title{parse out the forward and reverse UMIs and contig for a BED/BAM}
\usage{
parse_spike_UMI(UMI, pos = NULL, seqs = NULL)
}
\arguments{
\item{UMI}{a vector of UMIs}

\item{pos}{optional vector of positions (else all are set to 1)}

\item{seqs}{optional vector of read sequences (else widths default to 96)}
}
\value{
a GRanges
}
\description{
parse out the forward and reverse UMIs and contig for a BED/BAM
}
