% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedpe.R
\name{read_bedpe}
\alias{read_bedpe}
\title{read a BEDPE file into Pairs of GRanges (as if a GAlignmentPairs or similar)}
\usage{
read_bedpe(x, ..., stranded = FALSE, fraglen = TRUE)
}
\arguments{
\item{x}{a Tabixed BEDPE file, or a TabixFile of one}

\item{...}{additional arguments to pass to scanTabix internally}

\item{stranded}{is this BEDPE file stranded? (FALSE)}

\item{fraglen}{compute the fragment length? (TRUE)}
}
\value{
\preformatted{     a Pairs of GRanges, perhaps with $score or $fraglen
}
}
\description{
read a BEDPE file into Pairs of GRanges (as if a GAlignmentPairs or similar)
}
\details{
\preformatted{    BEDPE import in R is a shambles. This is a bandaid on a GSW.
             if all(score(pe) == 1), mcols(pe)$score will be dropped.
}
}
\examples{
\dontrun{
  bedpe <- "GSM5067076_2020_A64_bedpe.bed.gz"
  WT1_hg38 <- GRanges("chr11", IRanges(32387775, 32435564), "-")
  read_bedpe(bedpe, param=WT1_hg38)
  # note that since all(score(bedpe) == 1, score is dropped)
}

}
\seealso{
\preformatted{    bedpe_covg
}
}
