% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_spikes.R
\name{rename_spikes}
\alias{rename_spikes}
\title{for BAM/CRAM files with renamed contigs, we need to rename \code{spike} rows}
\usage{
rename_spikes(x, spike)
}
\arguments{
\item{x}{a BAM/CRAM file, hopefully with an index}

\item{spike}{a DataFrame where spike$sequence is a DNAStringSet}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a DataFrame with renamed contigs (rows)
}\if{html}{\out{</div>}}
}
\description{
This function does that.
}
\seealso{
generate_spike_fasta
}
