% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_spiked_bam.R
\name{scan_spiked_bam}
\alias{scan_spiked_bam}
\title{pretty much what it says: scan standard chroms + spike contigs from a BAM}
\usage{
scan_spiked_bam(
  bam,
  mapq = 20,
  binwidth = 300L,
  bins = NULL,
  spike = NULL,
  how = c("max", "mean"),
  dupe = FALSE,
  paired = TRUE,
  standard = TRUE,
  ...
)
}
\arguments{
\item{bam}{the BAM file}

\item{mapq}{minimum mapq value to count a pair (20)}

\item{binwidth}{width of the bins for chromosomal tiling (300)}

\item{bins}{a pre-tiled GRanges for binning coverage (NULL)}

\item{spike}{the spike-in references, if not using the defaults (NULL)}

\item{how}{how to record spike read coverage (max or mean)? (max)}

\item{dupe}{unique (FALSE), duplicte (TRUE), or all (NA) reads? (FALSE)}

\item{paired}{restrict coverage to that from properly paired reads? (TRUE)}

\item{standard}{restrict non-spike contigs to "standard" chromosomes? (TRUE)}

\item{...}{additional arguments to pass to scanBamFlag()}
}
\value{
\preformatted{     base-level read pair coverage on chroms + spikes in the BAM
}
}
\description{
pretty much what it says: scan standard chroms + spike contigs from a BAM
}
\details{
For example (not run), one might do something like:

data(spike, package="spiky")
bam <- "2021_ctl.hg38_withSpikes.bam" # a merged sorted BAM
spiked_covg <- scan_spiked_bam(bam, mapq=20, spikes=spike)
}
\seealso{
\preformatted{    GenomeInfoDb::keepStandardChromosomes
}

\preformatted{    Rsamtools::ScanBamParam
}
}
