% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike}
\alias{spike}
\title{spike-in contig properties for Sam's cfMeDIP spikes}
\format{
A DataFrame object with
\describe{
\item{sequence}{contig sequence, as a DNAStringSet}
\item{methylated}{are the CpGs in this spike-in methylated? 0 or 1}
\item{CpGs}{number of CpG dinucleotides in the spike, from 1 to 16}
\item{fmol}{femtomolar concentration of the spike-in for standard mix}
\item{molmass}{molar mass of spike-in sequence}
}
}
\usage{
spike
}
\description{
A DataFrame with sequence, concentration, and other properties of Sam's
synthetic cfMeDIP spike-in controls. The row names redudantly encode some
of these properties, such as the number of CpGs in the spike-in sequence.
}
\keyword{datasets}
