% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_bland_altman_plot.R
\name{spike_bland_altman_plot}
\alias{spike_bland_altman_plot}
\title{Bland-Altman plot for cfMeDIP spike standards}
\usage{
spike_bland_altman_plot(fit)
}
\arguments{
\item{fit}{a model fit, from predict_pmol (?)}
}
\value{
\preformatted{             a ggplot2 object
}
}
\description{
Bland-Altman plot for cfMeDIP spike standards
}
\examples{

data(spike_read_counts)
fit <- model_glm_pmol(spike_read_counts) 
spike_bland_altman_plot(fit) 

}
