% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike_cram_counts}
\alias{spike_cram_counts}
\title{spike-in counts, as a long data.frame}
\format{
A data.frame object with
\describe{
\item{frag_grp}{the encoded spike contig name: basepairs_CpGs_GCpercent}
\item{id}{subject from whom cfMeDIP spike reads (column 3) were counted}
\item{read_count}{read coverage for this spike in this subject (column 2)}
}
}
\source{
Generated from scan_spike_counts(CRAMs, spike=spike) using example CRAMs
containing spike contigs
}
\usage{
data(spike_cram_counts)
}
\description{
A data.frame with spike-in results from CRAM files (generated from scan_spike_counts(CRAMs, spike=spike))
}
\keyword{datasets}
