% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike_read_counts}
\alias{spike_read_counts}
\title{spike-in counts, as a long data.frame}
\format{
A data.frame object with
\describe{
\item{frag_grp}{the encoded spike contig name: basepairs_CpGs_GCpercent}
\item{id}{subject from whom cfMeDIP spike reads (column 3) were counted}
\item{read_count}{read coverage for this spike in this subject (column 2)}
}
}
\source{
This data was created using inst/script/loadDedup.R
}
\usage{
data(spike_read_counts)
}
\description{
A data.frame with spike-in results from control samples in the manuscript.
This maps 1:1 onto \code{dedup} using reshape2::melt.
}
\keyword{datasets}
