% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike_res}
\alias{spike_res}
\title{A Granges object with spike-in sequence coverage, and
summarized for each spike contig as (the default) \code{max} coverage. (In other
words, the default output of scan_spike_contigs or scan_spike_bedpe)
This represents what most users will want to generate from their own spike-in BAMs or BEDPEs, and
is used repeatedly in downstream examples throughout the package.}
\format{
A GRanges of coverage results with one metadata column, \code{coverage}
}
\source{
Generated using scan_spike_bedpe or scan_spike_contigs on an example bedpe or bam
containing spike contigs.
}
\usage{
data(spike_res)
}
\description{
A Granges object with spike-in sequence coverage, and
summarized for each spike contig as (the default) \code{max} coverage. (In other
words, the default output of scan_spike_contigs or scan_spike_bedpe)
This represents what most users will want to generate from their own spike-in BAMs or BEDPEs, and
is used repeatedly in downstream examples throughout the package.
}
\keyword{datasets}
