% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeBlandAltmanPlot.R
\name{spikeBlandAltmanPlot}
\alias{spikeBlandAltmanPlot}
\title{Bland-Altman plot for cfMeDIP spike standards}
\usage{
spikeBlandAltmanPlot(fit, ...)
}
\arguments{
\item{fit}{a model fit, from fitSpikeGlm}
}
\value{
\preformatted{             a ggplot2 object
}
}
\description{
Bland-Altman plot for cfMeDIP spike standards
}
