% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{spiky-methods}
\alias{spiky-methods}
\title{A handful of methods that I've always felt were missing}
\arguments{
\item{x}{an Rle or RleList, usually}
}
\description{
Particularly, simple methods to plot coverage results.
}
\details{
selectMethod("plot", "Rle") and also
selectMethod("plot", "RleList") too.
}
