% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssb_res}
\alias{ssb_res}
\title{scan_spiked_bam results from a merged cfMeDIP CRAM file (chr22 and spikes)}
\format{
A CompressedGRangesList of coverage results, containing
\describe{
\item{genomic}{a GRanges with one metadata column, \code{coverage}}
\item{spikes}{a GRanges with one metadata column, \code{coverage}}
}
}
\source{
Generated using scan_spiked_bam on an example bam
containing chr22 and spike contigs.
}
\usage{
data(ssb_res)
}
\description{
A CompressedGRangesList object with \code{genomic} (chr22) and \code{spikes} coverage,
binned every 300bp for the genomic contigs then averaged across the bin, and
summarized for each spike contig as (the default) \code{max} coverage. (In other
words, the default output of scan_spiked_bam, restricted to a small enough
set of genomic regions to be practical for examples.) This represents what
most users will want to generate from their own merged BAMs or CRAMs, and
is used repeatedly in downstream examples throughout the package.
}
\keyword{datasets}
