% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cram_support.R
\name{test_cram_support}
\alias{test_cram_support}
\title{a function to determine if CRAM support is enabled}
\usage{
test_cram_support(cram = NULL, reads = 1000L)
}
\arguments{
\item{cram}{a CRAM filename (default is example.spike.cram in extdata)}

\item{reads}{how many reads to fetch, i.e. the yieldSize (1000)}
}
\value{
TRUE or FALSE, perhaps with a message suggesting how to make it TRUE
}
\description{
At present, CRAM support is not enabled by default in Rsamtools. To fix this,
it may be necessary to use BiocManager::install() on GitHub branches, i.e.
BiocManager::install(c(
"Bioconductor/Rhtslib@enable-cram",
"Bioconductor/Rsamtools@cram"
))
}
\details{
Until such time as CRAM support is included by default, though, we need to
know whether access to CRAMs is available.  This function tests that.  If
the function detects that CRAM support is NOT installed, it instructs the
user on the best known way to enable it. (Ultimately, we want it in RELEASE)
}
\examples{
cram <- system.file("extdata", "example.spike.cram", package="spiky",
                  mustWork=TRUE)
\dontrun{
  test_cram_support(cram)
}

}
