% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testGR}
\alias{testGR}
\title{a test GRanges with UMI'ed genomic sequences used as controls}
\format{
A GRanges object with an mcols() DataFrame containing
\describe{
\item{UMI1}{the unique molecular identifier on the forward read}
\item{UMI2}{the unique molecular identifier on the reverse read}
\item{seq}{the sequence of the fragment}
\item{name}{the name of the fragment}
\item{score}{whether the fragment passes filters (always 1)}
}
}
\usage{
data(testGR)
}
\description{
Sources and overlap widths of various read sequences in a test CRAM.
}
\keyword{datasets}
