% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bedpe.R
\name{write_bedpe}
\alias{write_bedpe}
\title{write out a Pairs object as a BEDPE file}
\usage{
write_bedpe(object, con, ...)
}
\arguments{
\item{object}{a Pairs object (contrast with most export... functions)}

\item{con}{a connection (currently just a filename, see above)}
}
\value{
\preformatted{   the results of write.table(..., con)
}
}
\description{
see read_bedpe for more details on the BEDPE format itself; note that this
could eventually be patched into rtracklayer, but don't hold your breath.
}
\details{
\preformatted{  Unlike read_bedpe, write_bedpe tries to respect extra columns
}
}
