% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pmol.R
\name{bin_pmol}
\alias{bin_pmol}
\title{Binned estimation of picomoles of DNA present in cfMeDIP assays}
\usage{
bin_pmol(x)
}
\arguments{
\item{x}{results from predict_pmol (a data.frame)}
}
\value{
\preformatted{ the same data.frame, but with a column `adjusted_pred_con`
}
}
\description{
Given the results of model_glm_pmol and predict_pmol, adjust the predictions
to reflect picomoles of captured DNA overlapping a given bin in the genome.
}
\seealso{
model_glm_pmol

predict_pmol
}
