% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pmol.R
\name{bin_pmol}
\alias{bin_pmol}
\title{This function replaces the following line in bedtools:}
\usage{
bin_pmol(...)
}
\description{
bedtools intersect -wao -a fragments.bed -b hg38_300bp_windows.bed > data.bed
}
\details{
Which translates to "Write the original A and B entries,
plus the number of base pairs of overlap between the two features".

Which translates to
}
