% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spike_fasta.R
\name{generate_spike_fasta}
\alias{generate_spike_fasta}
\title{for CRAM files, a FASTA reference is required to decode; this builds that}
\usage{
generate_spike_fasta(x, assembly = NULL, fasta = "spike_contigs.fa")
}
\arguments{
\item{x}{a CRAM file, hopefully with an index}

\item{assembly}{optional BSgenome or seqinfo with reference contigs (NULL)}

\item{fasta}{the filename for the resulting FASTA ("spikes.fa")}
}
\value{
\preformatted{     invisibly, a DNAStringSet as exported to `fasta` 
}
}
\description{
If the contigs in a CRAM have even slightly different names from those in
the reference, decoding will fail.  In some cases there are multiple names
for a given contig (which raises the question of whether to condense them),
and thus the same reference sequence decodes multiple contig names.
}
\details{
This function simply generates an appropriate spike reference for a CRAM,
using the CRAM headers to figure out which references are used for which.
At the moment, CRAM support in Rsamtools only exists in the GitHub branch:

BiocManager::install("Bioconductor/Rsamtools@cram")

Using other versions of Rsamtools will yield an error on CRAM files.

Note that for merged genomic + spike reference BAMs/CRAMs, this function
will only attempt to generate a FASTA for the spike contigs, not reference.
If your reference contigs are screwed up, talk to your sequencing people,
and keep better track of the FASTA reference against which you compress!
}
\seealso{
\preformatted{    rename_contigs
}
}
