% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binned_coverage.R
\name{get_binned_coverage}
\alias{get_binned_coverage}
\title{tabulate read coverage in predefined bins}
\usage{
get_binned_coverage(bins, covg)
}
\arguments{
\item{bins}{the GRanges with bins}

\item{covg}{the coverage result (an RleList)}
}
\value{
\preformatted{   a GRanges of summarized coverage
}
}
\description{
refactored out of scan_spiked_bam
}
\seealso{
\preformatted{  get_spiked_coverage
}

\preformatted{  scan_spiked_bam
}
}
