% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merged_gr.R
\name{get_merged_gr}
\alias{get_merged_gr}
\title{get a GRanges of (by default, standard) chromosomes from seqinfo}
\usage{
get_merged_gr(si, standard = TRUE)
}
\arguments{
\item{si}{seqinfo, usually from a BAM/CRAM file with spike contigs}

\item{standard}{trim to standard chromosomes? (TRUE)}
}
\value{
\preformatted{     GRanges with two genomes: the organism assembly and "spike"
}
}
\description{
refactored from scan_spiked_bam to clarify information flow
}
