% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_specificity.R
\name{methylation_specificity}
\alias{methylation_specificity}
\title{compute methylation specificity for spike-in standards}
\usage{
methylation_specificity(ssb_res, spike)
}
\arguments{
\item{ssb_res}{output object from scan_spiked_bam}

\item{spike}{spike contig database, if needed (e.g. data(spike))}
}
\value{
\preformatted{         list with median and mean coverage across spike contigs
}
}
\description{
In a cfMeDIP experiment, the yield of methylated fragments should be >95\%
(ideally 98-99\%) due to the nature of the assay.
}
\examples{
data(ssb_res)
data(spike, package="spiky")
methylation_specificity(ssb_res,spike=spike)

}
