% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pmol.R
\name{predict_pmol}
\alias{predict_pmol}
\title{predict picomoles of DNA from a fit and readcounts}
\usage{
predict_pmol(fit, x = NULL)
}
\arguments{
\item{fit}{result of model_glm_pmol}

\item{x}{the data / new data}
}
\value{
x, but with predictions
}
\description{
this seems to just wrap output from bin_pmol and model_glm_pmol... ?
}
\examples{

data(spike_read_counts)
fit <- model_glm_pmol(spike_read_counts) 
pred <- predict_pmol(fit) 

}
