% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spikes.R
\name{process_spikes}
\alias{process_spikes}
\title{QC, QA, and processing for a new spike database}
\usage{
process_spikes(fasta, methylated = 0, ...)
}
\arguments{
\item{fasta}{fasta file with new spike sequences}

\item{methylated}{whether CpGs in each are methylated (0 or 1, default 0)}

\item{...}{additional arguments, e.g. kernels (currently unused)}
}
\value{
\preformatted{         a DataFrame suitable for downstream processing 
}
}
\description{
Never trust anyone, least of all yourself
}
\details{
GCfrac is the GC content of spikes as a proportion instead of a percent.
OECpG is (observed/expected) CpGs (expectation is 25\% of GC dinucleotides).
}
\examples{

phages <- system.file("extdata", "phages.fa", package="spiky", mustWork=TRUE)
process_spikes(phages)

data(spike)
spikes <- system.file("extdata", "spikes.fa", package="spiky", mustWork=TRUE)
spikemeth <- spike$methylated
process_spikes(spikes, spikemeth)

# not run
# library(kebabs)
# CpGmotifs <- c("[AT]CG[AT]","C[ATC]G", "CCGG", "CGCG")
# mot <- motifKernel(CpGmotifs, normalized=FALSE)
# km <- getKernelMatrix(mot, subset(phages, methylated == 0)$sequence)
# heatmap(km, symm=TRUE)

}
