% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bedpe.R
\name{read_bedpe}
\alias{read_bedpe}
\title{read a BEDPE file into Pairs of GRanges (as if a GAlignmentPairs or similar)}
\usage{
read_bedpe(x, param = NULL, stranded = TRUE)
}
\arguments{
\item{x}{a Tabixed BEDPE file, or a TabixFile of one}

\item{param}{a GRanges of regions to read, or NULL (NULL)}

\item{stranded}{retain strand of GRanges in the resulting Pairs? (TRUE)}
}
\value{
\preformatted{     an S4Vectors::Pairs of GRanges with $score in mcols() 
}
}
\description{
read a BEDPE file into Pairs of GRanges (as if a GAlignmentPairs or similar)
}
\details{
\preformatted{    BEDPE import in R is a shambles. This is a bandaid on a GSW.
}
}
\seealso{
\preformatted{    bedpe_covg
}
}
