% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_spike_seqlevels.R
\name{rename_spike_seqlevels}
\alias{rename_spike_seqlevels}
\title{for spike-in contigs in GRanges, match to standardized spike seqlevels}
\usage{
rename_spike_seqlevels(x, spike)
}
\arguments{
\item{x}{something with seqlevels}

\item{spike}{a DataFrame where spike$sequence is a DNAStringSet}
}
\value{
\preformatted{     x, but with standardized spike seqlevels
}
}
\description{
This function is essentially the opposite of rename_spikes, except that
it works well on GRanges/GAlignments from or for merged genome+spike BAMs.
}
\seealso{
\preformatted{    rename_spikes
}
}
