% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_bland_altman_plot.R
\name{spike_bland_altman_plot}
\alias{spike_bland_altman_plot}
\title{Bland-Altman plot for cfMeDIP spike standards}
\usage{
spike_bland_altman_plot(fit)
}
\arguments{
\item{fit}{a model fit, from predict_pmol (?)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{             a ggplot2 object
}\if{html}{\out{</div>}}
}
\description{
Bland-Altman plot for cfMeDIP spike standards
}
\examples{

data(spike_res)
data(spike, package="spiky")
fit <- model_glm_pmol(covg_to_df(spike_res, spike=spike),spike=spike)
ba_plot <- spike_bland_altman_plot(fit)

}
